+++ src/main/app/src/entities/AsennuspaikanTyyppi.tsx
